updated project to use flask blueprints for better structure
