+++ calculapiz/app/config.py
@@ -1,0 +1,4 @@
+import os
+
+DATA_PATH = 'data'
+STATISTICS_PATH = os.path.join(DATA_PATH, 'statistics.csv')
