+++ calculapiz/app/operations/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('operations', __name__)
+
+from app.operations import routes
