+++ calculapiz/app/operations/operations.py
@@ -1,0 +1,9 @@
+"""This module contains enumeration for operations available in service"""
+from enum import Enum
+
+
+class Operations(Enum):
+    ADD = 'add'
+    SUB = 'sub'
+    MLP = 'mlp'
+    DIV = 'div'
