+++ calculapiz/app/templates/about.html
@@ -6,7 +6,7 @@
     </head>
 
     <body>
-        <h1>Calculapiz v0.0.1</h1>
+        <h1>Calculapiz v0.0.2</h1>
 
         <div>
             <h2>Description</h2>
