--- calculapiz/requirements.txt
@@ -1,1 +1,0 @@
-Flask==3.0.2
